Moved functions from anno.py and scoring.py to utils
Re-organization of module structure for anno.py and scoring.py
